ギャラリーページのカスタムフック作成
