+++ src/app/gallery/_hooks/useControlGalleryImage.ts
@@ -1,0 +1,339 @@
+"use client";
+import { ChangeEvent, useCallback, useEffect, useRef, useState } from "react";
+import toast from "react-hot-toast";
+import { useSupabaseSession } from "@/app/_hooks/useSupabaseSession";
+import { GalleryItem } from "@/app/_type/Gallery";
+import { v4 as uuidv4 } from "uuid";
+import { supabase } from "@/utils/supabase";
+
+const useControlGalleryImage = (selectedTabId: number) => {
+  const { token } = useSupabaseSession();
+  const [isLoading, setIsLoading] = useState(false);
+  const [isImgModalOpen, setIsImgModalOpen] = useState(false);
+  const [thumbnailImageKey, setThumbnailImageKey] = useState<string>("");
+  const fileInputRef = useRef<HTMLInputElement | null>(null);
+  const [thumbnailImageUrls, setThumbnailImageUrls] = useState<GalleryItem[]>(
+    []
+  );
+  const [selectedImageUrl, setSelectedImageUrl] = useState<string | null>(null); // クリックされた画像をModal表示（URLを保持する状態）
+  const [selectedImageId, setSelectedImageId] = useState<number>();
+
+  const generateSignedImageUrl = async (key: string) => {
+    const { data, error } = await supabase.storage
+      .from("gallery_item") // 正しいバケット名を指定
+      .createSignedUrl(key, 60 * 60); // 有効期限を1時間に設定
+
+    if (error) {
+      console.error("Error creating signed URL:", error.message);
+      return null;
+    }
+    return data.signedUrl;
+  };
+
+  // 画像のURLを取得するためのuseEffect;
+  const fetchGalleryItems = useCallback(async () => {
+    if (!token || !selectedTabId) return;
+
+    setIsLoading(true);
+    try {
+      // API経由でGalleryItemsテーブルからデータを取得
+      const response = await fetch(
+        `/api/gallery_group/${selectedTabId}/gallery_items`,
+        {
+          headers: {
+            Authorization: token!,
+          },
+        }
+      );
+
+      const data = await response.json();
+      if (!response.ok || !data.galleryItems) {
+        console.error("Failed to fetch gallery items:", data);
+        setThumbnailImageUrls([]);
+        return;
+      }
+
+      const itemsWithUrls: GalleryItem[] = await Promise.all(
+        data.galleryItems.map(async (item: GalleryItem) => {
+          const signedUrl = await generateSignedImageUrl(
+            item.thumbnailImageKey
+          );
+          return { ...item, signedUrl }; // GalleryItemにsignedUrlを追加
+        })
+      );
+
+      // 有効なURLのみをセット
+      setThumbnailImageUrls(
+        itemsWithUrls.filter((item) => item.signedUrl !== null)
+      );
+    } catch (error) {
+      console.error("Error fetching gallery items:", error);
+      setThumbnailImageUrls([]);
+    } finally {
+      setIsLoading(false);
+    }
+  }, [token, selectedTabId]);
+
+  useEffect(() => {
+    fetchGalleryItems();
+  }, [fetchGalleryItems, selectedTabId]); // useCallback でメモ化された fetchGalleryItems を依存関係に追加
+  //画像をstorageに追加
+  const handleImageChange = async (event: ChangeEvent<HTMLInputElement>) => {
+    try {
+      if (!event.target.files || event.target.files.length === 0) {
+        console.error("No image file selected.");
+        return;
+      }
+
+      const file = event.target.files[0];
+      const filePath = `private/${uuidv4()}`;
+
+      // Supabase ストレージに画像をアップロード
+      let uploadData;
+      try {
+        const { data, error } = await supabase.storage
+          .from("gallery_item")
+          .upload(filePath, file, {
+            cacheControl: "3600",
+            upsert: false,
+          });
+
+        if (error) {
+          throw new Error(`Upload error: ${error.message}`);
+        }
+
+        uploadData = data;
+        setThumbnailImageKey(uploadData.path);
+      } catch (uploadError) {
+        console.error("Error uploading image to Supabase:", uploadError);
+        alert("画像のアップロードに失敗しました。");
+        return;
+      }
+
+      // API に画像情報を送信
+      try {
+        const response = await fetch(
+          `/api/gallery_group/${selectedTabId}/gallery_items`,
+          {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+              Authorization: token!,
+            },
+            body: JSON.stringify({
+              galleryGroupId: selectedTabId,
+              thumbnailImageKey: uploadData.path,
+            }),
+          }
+        );
+
+        if (response.ok) {
+          fetchGalleryItems(); // 画像のリストを再取得
+          toast.success("画像が正常に追加されました。", {
+            duration: 2100,
+          });
+        } else {
+          const errorData = await response.json();
+          console.error("Failed to add image to gallery:", errorData);
+          throw new Error(
+            `API error: ${errorData.message || "Unknown error occurred"}`
+          );
+        }
+      } catch (apiError) {
+        console.error("Error adding image to gallery:", apiError);
+        alert("画像の追加に失敗しました。");
+      }
+    } catch (generalError) {
+      console.error("Unexpected error occurred:", generalError);
+      alert("予期しないエラーが発生しました。もう一度お試しください。");
+    }
+  };
+
+  //ファイルを一時的にアクセス可能にする サインドURL
+  const fetchSignedUrl = async (thumbnailImageKey: string) => {
+    try {
+      const { data, error } = await supabase.storage
+        .from("gallery_item") // バケット名を指定
+        .createSignedUrl(thumbnailImageKey, 60 * 60); // 有効期限を1時間(3600秒)に設定
+
+      if (error) {
+        console.error("Error creating signed URL:", error.message);
+        return null;
+      }
+
+      return data.signedUrl; // サインドURLを返す
+    } catch (error) {
+      console.error("Error fetching signed URL:", error);
+      return null;
+    }
+  };
+  useEffect(() => {
+    if (!token || !selectedTabId || !thumbnailImageKey) return;
+    const fetcher = async () => {
+      const signedUrl = await fetchSignedUrl(thumbnailImageKey);
+      if (signedUrl) {
+        setSelectedImageUrl(signedUrl); // 単一のURLを設定
+      }
+    };
+    fetcher();
+  }, [thumbnailImageKey, token, selectedTabId]);
+
+  //画像拡大Modal
+
+  // 画像クリック時の処理
+  const handleImgClick = (url: string, key: string, id: number) => {
+    setSelectedImageUrl(url);
+    setThumbnailImageKey(key); // 正しいキーを設定
+    setSelectedImageId(id); // 正しいIDを設定
+    setIsImgModalOpen(true);
+  };
+
+  // モーダルを閉じる処理を修正して画像リセット
+  const closeImgModal = () => {
+    setIsImgModalOpen(false);
+    setSelectedImageUrl(null); // 選択された画像をリセット;
+  };
+
+  const handleAddEvent = () => {
+    if (fileInputRef.current) {
+      fileInputRef.current.click(); // inputのクリックイベントをトリガー
+    }
+  };
+
+  //画像変更（Modal）
+  const updateImg = async (id?: number) => {
+    if (!fileInputRef.current || !fileInputRef.current.files) {
+      console.error("File input is not available or no file selected.");
+      return;
+    }
+
+    const file = fileInputRef.current.files[0];
+    if (!file) {
+      console.error("No image file selected.");
+      return;
+    }
+
+    const newImageKey = `private/${uuidv4()}`;
+    try {
+      const { error: uploadError } = await supabase.storage
+        .from("gallery_item")
+        .upload(newImageKey, file, {
+          cacheControl: "3600",
+          upsert: true,
+        });
+
+      if (uploadError) {
+        console.error("Image upload failed:", uploadError.message);
+        return;
+      }
+      // 古い画像を削除する処理
+      const { error: deleteError } = await supabase.storage
+        .from("gallery_item")
+        .remove([thumbnailImageKey]);
+
+      if (deleteError) {
+        console.error("Failed to delete old image:", deleteError.message);
+      }
+      // API 経由での更新処理
+      const response = await fetch(
+        `/api/gallery_group/${selectedTabId}/gallery_items/${id}`,
+        {
+          method: "PUT",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token!,
+          },
+          body: JSON.stringify({
+            galleryGroupId: selectedTabId,
+            thumbnailImageKey: newImageKey,
+          }),
+        }
+      );
+
+      if (response.ok) {
+        await fetchGalleryItems();
+
+        toast.success("画像が変更されました。", {
+          duration: 3000, //ポップアップ表示時間
+        });
+        closeImgModal();
+      } else {
+        console.error(
+          "Failed to update image in database:",
+          await response.json()
+        );
+      }
+    } catch (error) {
+      console.error("Error during image update:", error);
+    }
+  };
+  //画像削除（Modal）
+  const deleteImg = async (id?: number) => {
+    if (!token) return;
+    if (!selectedImageUrl || !thumbnailImageKey || id === undefined) {
+      console.error("No image selected for deletion.");
+      return;
+    }
+    if (!confirm("画像を削除しますか？")) return;
+    try {
+      // バケット内の画像を削除
+      const { error: storageError } = await supabase.storage
+        .from("gallery_item") // バケット名を指定
+        .remove([thumbnailImageKey]); // thumbnailImageKeyを利用
+
+      if (storageError) {
+        console.error(
+          "Failed to delete image from bucket:",
+          storageError.message
+        );
+        return;
+      }
+
+      const response = await fetch(
+        `/api/gallery_group/${selectedTabId}/gallery_items/${id}`,
+        {
+          method: "DELETE",
+          headers: {
+            "Content-Type": "application/json",
+            Authorization: token!,
+          },
+        }
+      );
+
+      if (response.ok) {
+        fetchGalleryItems(); // 最新状態を取得
+        // setSelectedImageUrl(null);
+        setThumbnailImageKey(""); // キーをリセット
+        closeImgModal();
+        toast.success("画像を一つ削除しました。", {
+          duration: 7000, //ポップアップ表示時間
+        });
+      } else {
+        console.error("Failed to delete tab");
+      }
+    } catch (error) {
+      console.error("Error deleting Image:", error);
+    }
+  };
+
+  return {
+    selectedTabId,
+    fileInputRef,
+    handleImageChange,
+    thumbnailImageUrls,
+    handleImgClick,
+    isImgModalOpen,
+    closeImgModal,
+    selectedImageUrl,
+    updateImg,
+    selectedImageId,
+    deleteImg,
+    generateSignedImageUrl,
+    fetchGalleryItems,
+    fetchSignedUrl,
+    handleAddEvent,
+    isLoading,
+  };
+};
+
+export default useControlGalleryImage;
